Refactor java test to assert from app code
