+++ example/android/javalib/1-hello-world/app/src/main/java/com/helloworld/SampleLogic.java
@@ -1,0 +1,9 @@
+package com.helloworld;
+
+
+class SampleLogic {
+
+    public static float textSize() {
+        return 32f;
+    }
+}
